python-backend/requirements.txt
# fastapi>=0.115.0
# uvicorn[standard]>=0.30.0

# langchain>=0.3.0
# langchain-core>=0.3.0
# langchain-community>=0.3.0
# langchain-groq>=0.3.0
# langchain-text-splitters>=0.3.0

# faiss-cpu>=1.8.0
# sentence-transformers>=3.0.0

# python-dotenv>=1.0.0
# pydantic>=2.7.0
# pydantic-settings>=2.2.0



# FastAPI server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# LangChain ecosystem
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-groq>=0.3.0
langchain-text-splitters>=0.3.0

# Vector store + embeddings (RAG)
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
httpx>=0.27.0

# Crop Monetizer
pandas>=2.0.0
numpy>=1.26.0

# Feed feature
PyJWT>=2.8.0
ddgs>=1.0.0